implement model recalculation on slider change
